Move All Commands To botFuncs Folder
- All commands have been moved to the botFuncs folder according to their purposes.
- Dependencies updated.

**change severity : high**
